berkelium: cmake build updates, refactoring

+++ build/clean.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
 cat .gitignore | while read line; do
+    [ "$line" == "eclipse" ] && continue
     rm -rf $line
 done
